requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.1.1
Flask_Migrate==4.0.5
Werkzeug==2.3.7
Jinja2==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
